# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Resource Manager
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /device:
    # binds a127 app logic to a route
    x-swagger-router-controller: device
    get:
      description: Returns list of registered devices
      # used as the method name of the controller
      operationId: getdevices
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetDeviceResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Registers a device to RM
      # used as the method name of the controller
      operationId: register
      parameters:
        - name: title
          in: body
          description: Device properties
          required: true
          schema:
            $ref: "#/definitions/Device"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/{id}:
    x-swagger-router-controller: device
    get:
      description: Get device application information
      operationId: getdevice
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDeviceAppResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    post:
      description: Save application in the device
      operationId: saveapp
      parameters:
        - name: id
          description: Device ID
          type: string
          in: path
          required: true
        - name: title
          description: Application properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Application"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/{id}/{app_id}:
    x-swagger-router-controller: device
    put:
      description: Update application in the device
      operationId: updateapp
      parameters:
        - name: id
          description: Device ID
          type: string
          in: path
          required: true
        - name: app_id
          description: App ID
          type: string
          in: path
          required: true
        - name: title
          description: Application properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Application"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete given application from the device
      operationId: deleteapp
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: app_id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GeneralResponse:
    type: object
    properties:
      success:
        type: string
        description: returns device or app ID
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  GetDeviceResponse:
    type: object
    required:
      - devices
      - message
    properties:
      message:
        type: string
      devices:
        type: array
        items:
          type: object
        properties:
          id: 
            type: string
          location: 
            type: string
          device_type: 
            type: array
            items:
              type: string
  GetDeviceAppResponse:
    required:
      - device
      - message
    type: object
    properties:
      message:
        type: string
      device:
        type: object
        properties:
          id:
            type: string
          location: 
            type: string
          device_type: 
            type: array
            items:
              type: string
          app:
            type: array
            items:
              type: object
            properties:
              id: 
                type: string
              name: 
                type: string
              version: 
                type: string 
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Device:
    type: object
    properties:
      # id:
      #   type: string
      #   description: Device ID
      location:
        type: string
        description: Device Location
      device_type:
        type: array
        items:
          type: string
        description: Device Type or Device Capability
    required:
      # - id
      - location
      - device_type
  Application:
    type: array
    items:
      type: object
      properties:
        # id:
        #   type: string
        #   description: Application ID
        image:
          type: string
          description: Application name
        # version:
        #   type: string
          # description: Application version
      required:
        # - id
        - image
        # - version